+++ employee-owner-login/Services/IInventoryService.cs
@@ -5,5 +5,5 @@ namespace PoultryOS.Services;
 public interface IInventoryService
 {
     DashboardSummaryViewModel GetSummary(string username, string role);
-    DashboardSummaryViewModel RecordSale(int productId, int quantity, string username, string role);
+    DashboardSummaryViewModel RecordSale(int productId, decimal quantity, string username, string role);
 }
